requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
cloudinary==1.26.0
defusedxml==0.7.1
Django==3.2.8
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
djangorestframework==3.12.4
drf-spectacular==0.20.2
httpie==2.6.0
idna==3.3
inflection==0.5.1
jsonschema==4.1.2
Pillow==8.4.0
psycopg2-binary==2.9.1
Pygments==2.10.0
pyrsistent==0.18.0
PySocks==1.7.1
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
routers==0.10.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
